% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coselenss.R
\name{coselenss}
\alias{coselenss}
\title{Calculate dN/dS in each of the groups (patients in the cancer_type group1 mutations in the split_gene and those group2).
Perform a log-likelihood ratio test on the log likelihoods obtained from running the groups individually and then together.
Save the p-value and data for later use.}
\usage{
coselenss(group1, group2, subset.genes.by)
}
\arguments{
\item{group1}{group of individuals (for example those that contain a mutation in a split_gene)}

\item{group2}{another group of individuals that do NOT contain a mutation in a split_gene}

\item{subset.genes.by}{genes to subset results by}
}
\description{
Calculate dN/dS in each of the groups (patients in the cancer_type group1 mutations in the split_gene and those group2).
Perform a log-likelihood ratio test on the log likelihoods obtained from running the groups individually and then together.
Save the p-value and data for later use.
}
