% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coselens.R
\name{coselens}
\alias{coselens}
\title{COnditional SELection on the Excess of NonSynonymous Substitutions (coselens)}
\usage{
coselens(
  group1,
  group2,
  subset.genes.by = NULL,
  sequenced.genes = NULL,
  refdb = "hg19",
  sm = "192r_3w",
  kc = "cgc81",
  cv = "hg19",
  max_muts_per_gene_per_sample = Inf,
  max_coding_muts_per_sample = 3000,
  use_indel_sites = T,
  min_indels = 5,
  maxcovs = 20,
  constrain_wnon_wspl = T,
  outp = 3,
  numcode = 1
)
}
\arguments{
\item{group1}{group of individuals (for example those that contain a mutation in a split_gene)}

\item{group2}{another group of individuals that do NOT contain a mutation in a split_gene}

\item{subset.genes.by}{genes to subset results by}

\item{sequenced.genes}{the gene_list paramater from dndscv, which is a list of genes to restrict the analysis (use for targeted sequencing studies)}

\item{...}{other parameters passed to dncdscv, all defaults from dndscv are used except max_muts_per_gene_per_sample is set to Infinity}
}
\value{
coselens returns a list containing four dataframes: 1) "summary" a summary of coselens output that should be sufficient for most users, 2) "full" which includes more detailed output, 3) fitted substitution models for group 1 from dndscv, and 4) fitted substitution models for group 2 from dndscv. The "summary" dataframe contains the following:

\itemize{
\item gene_name: name of gene that conditional selection was calculated in
}

\itemize{
\item num.driver.sub.group1: estimate of the number of drivers in group 1 based excess of non-synonymous mutations
}

\itemize{
\item num.driver.sub.group2: estimate of the number of drivers in group 2 based excess of non-synonymous mutations
}

\itemize{
\item num.driver.ind.group1: estimate of the number of drivers in group 1 based excess of indels
}

\itemize{
\item num.driver.ind.group2: estimate of the number of drivers in group 2 based excess of indels
}

\itemize{
\item psub: p-value for conditional selection in non-synonymous substitutions
}

\itemize{
\item pind: p-value for conditional selection in indels
}

\itemize{
\item pglobal: Fisher's combined p-value for psub and pind
}

\itemize{
\item qsub: q-value of psub using Benjamini-Hochberg correction
}

\itemize{
\item qind: q-value of pind using Benjamini-Hochberg correction
}

\itemize{
\item qglobal: q-value of pglobal using Benjamini-Hochberg correction
}

The "full" dataframe contains all the informtion from the "summary" dataframe, with the following additional columns:

\itemize{
\item num.driver.mis.group1: estimate of the number of drivers in group 1 based excess of missense mutations
}

\itemize{
\item num.driver.mis.group2: estimate of the number of drivers in group 2 based excess of missense mutations
}

\itemize{
\item num.driver.trunc.group1: estimate of the number of drivers in group 1 based excess of truncating mutations
}

\itemize{
\item num.driver.trunc.group2: estimate of the number of drivers in group 2 based excess of truncating mutations
}

\itemize{
\item pmis: p-value for conditional selection in missense mutations
}

\itemize{
\item ptrunc: p-value for conditional selection in truncating mutations
}

\itemize{
\item psub.group1: p-value for selection in group 1 for non-synonymous substitutions
}

\itemize{
\item psub.group2: p-value for selection in group 2 for non-synonymous substitutions
}

\itemize{
\item pmis.group1: p-value for selection in group 1 for missense substitutions
}

\itemize{
\item pmis.group2: p-value for selection in group 2 for missense substitutions
}

\itemize{
\item ptrunc.group1: p-value for selection in group 1 for truncating substitutions
}

\itemize{
\item ptrunc.group2: p-value for selection in group 2 for truncating substitutions
}

\itemize{
\item pind.group1: p-value for selection in group 1 for indels
}

\itemize{
\item pind.group2: p-value for selection in group 2 for indels
}

\itemize{
\item qsub.group1: q-value of psub.group1 using Benjamini-Hochberg correction
}

\itemize{
\item qsub.group2: q-value of psub.group2 using Benjamini-Hochberg correction
}

\itemize{
\item qmis.group1: q-value of pmis.group1 using Benjamini-Hochberg correction
}

\itemize{
\item qmis.group2: q-value of pmis.group2 using Benjamini-Hochberg correction
}

\itemize{
\item qtrunc.group1: q-value of ptrunc.group1 using Benjamini-Hochberg correction
}

\itemize{
\item qtrunc.group2: q-value of ptrunc.group2 using Benjamini-Hochberg correction
}

\itemize{
\item qind.group1: q-value of pind.group1 using Benjamini-Hochberg correction
}

\itemize{
\item qind.group2: q-value of pind.group2 using Benjamini-Hochberg correction
}
}
\description{
coselens identifies genes that are differentially selected depending on the grouping variable and provides maximum likelihood estimates of the effect sizes.
Given an input of mutations from two groups of patients, dN/dS is calculated in each of the groups,
and the excess of non-synonymous substitutions between them is found as an estimate for the difference
in the number of driver mutations. For example, one group can be mutations from individuals in a certain
cancer type with mutations in a specific gene vs those without mutations in the specific gene.
A p value is returned for differential selection in each gene in the reference genome (default: hg19).
}
