% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coselenss.R
\name{coselenss}
\alias{coselenss}
\title{COnditional SELection on the Excess of NonSynonymous Substitutions (COSELENSS)}
\usage{
coselenss(group1, group2, subset.genes.by = NULL)
}
\arguments{
\item{group1}{group of individuals (for example those that contain a mutation in a split_gene)}

\item{group2}{another group of individuals that do NOT contain a mutation in a split_gene}

\item{subset.genes.by}{genes to subset results by}
}
\value{
coselenss returns a dataframe with rows representing and the following columns

\itemize{
\item column1: column 1 description
}

\itemize{
\item column2: column 2 description
}
}
\description{
Given an input of mutations from two groups of patients, calculate dN/dS in each of the groups,
and find the excess of non-synonymous substitutions between them. For example, one group can be
mutations from individuals in a certain cancer type with mutations in a specific gene vs
those without mutations in the specific gene. A p-value is returned.
}
